Added method to check existing files and XL data

+++ download_codes.py
@@ -4,6 +4,7 @@ import time
 from zipfile import ZipFile
 import concurrent.futures
 import logging
+import os
 
 
 logging.basicConfig(filename="logs.log", 
@@ -19,31 +20,33 @@ def unzip_files(file_name, date, name):
     try:
         # opening the zip file in READ mode 
         with ZipFile(file_name, 'r') as zip:    
-            zip.extractall(path = f'Downloads\\Contest\\{date}\\{name}//')   # extracting all the files 
+            zip.extractall(path = f'Downloads\\Non Contest\\{date}\\{name}//')   # extracting all the files 
 
     except Exception as e:
-        # print(f"{name} not valid file")
         logger.warn(f"{name} not valid zip file") 
 
 
 def download_files(name, date, link):
     '''Dowload and save zip files to a folder from given link'''
-
-    try:
-        print("downloading {}".format(name))
-        data = requests.get(link).content
-        file_name = f'Downloads\\Contest\\Zip\\{name}.zip'
-        with open(file_name, 'wb') as _file:
-            _file.write(data)
+    
+    file_name = f'Downloads\\Non Contest\\Zip\\{name}.zip'
+    if not os.path.isfile(file_name):
+        try:
+            data = requests.get(link).content
             
-        unzip_files(file_name, date, name)
-        return f'{name} was downloaded and unzipped...'
-    except Exception as e:
-        logger.warn(f"{name} not downloadable") 
+            with open(file_name, 'wb') as _file:
+                _file.write(data)
+                
+            unzip_files(file_name, date, name)
+            return f'{name} was downloaded and unzipped...'
+        except Exception as e:
+            logger.warn(f"{name} not downloadable") 
+    else:
+        print(f"{name} already exists")
 
 path = "E:\Codes\Python\Web-Automate-Selenium\Internshala-Python-Users.xlsx"  #XL sheet path
 book = openpyxl.load_workbook(path)
-sheet = book.get_sheet_by_name("contest_users")   #fetch all user data
+sheet = book.get_sheet_by_name("non_contest_users")   #fetch all user data
 
 names = [name.value for name in sheet["A"][1:]]
 dates = [date.value.split(" ")[0] for date in sheet["C"][1:]]
